Allow to override the location of Cassandra's data (#19)
Before, the path was hard-coded, which gave no flexibility and the
outer system had to be adjusted to the role's behaviour. The change
allows to take the opposite perspective, that is to adjust the
Cassandra installation to the environment-specific needs.

Please note that this change is fully backward-compatible, because
the value defaults to the previous hard-coded one.